+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 -   Added [`zod`](https://zod.dev/) as an importable library for user scripts.
 -   Improved `os.installPackage()` to support upgrading and downgrading packages.
     -   When downgrading a package, the `downgrade` property on the `options` parameter object should be set to `true`. This is a safety feature to help prevent accidentally downgrading a package version.
+-   Added the `automaticBiosOptionInst` web config option.
+    -   When set to an inst ID with the `automaticBiosOption`, CasualOS will automatically load the specified inst instead of generating a random one.
+    -   inst IDs are formatted like this: `{recordName}/{inst}`. For public (free) or static (local) insts, omit the record name: `/{inst}`.
 -   Added the ability to "login to a studio" via custom domains.
     -   Users will receive a new account when logging into CasualOS served through a custom domain that is scoped to the studio that owns the custom domain.
     -   This provides an additional privacy boundary for users, helping ensure that data isn't automatically shared with custom domains simply by logging in.